requirements.txt
torch
torchvision
torchaudio
opencv-python
matplotlib
pillow
fastapi
uvicorn
